requirements/base.txt
asn1crypto==0.24.0
boto3==1.9.236
botocore==1.12.236
cffi==1.12.3
cryptography==2.7
Django==2.2.5
django-cors-headers==3.1.0
django-extensions==2.2.1
django-rest-knox==4.1.0
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-recursive==0.1.2
docutils==0.15.2
jmespath==0.9.4
Pillow==6.1.0
pycparser==2.19
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
